pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-backtester"
version = "0.1.0"
description = "A simple but powerful cryptocurrency backtesting engine"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ccxt>=4.0.0",
    "backtrader>=1.9.78.123",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "ta>=0.11.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "psutil>=5.9.0",
    "APScheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

